Started working on model viewer. Have issues.
